+++ tasks/gutyansky_a_matrix_column_sum/common/include/vector.hpp
@@ -7,7 +7,7 @@ namespace gutyansky_a_matrix_column_sum {
 
 struct Vector {
   size_t size;
-  std::vector<double> data;
+  std::vector<int64_t> data;
 
   friend bool operator==(const Vector &v_left, const Vector &v_right) {
     return v_left.size == v_right.size && v_left.data == v_right.data;
